[log] Extract `trace_loc` utility
Move `trace_loc` to its own header, to make it available to other code.
